Add AgeVerification program for voting eligibility
The Age Verification program asks the user to enter their age. It then uses an if-else statement to check whether the age is 18 or above. If the age is 18 or older, the program displays a message stating that the user is eligible to vote. Otherwise, it displays a message stating that the user is not eligible to vote.