Create VendingMachine.java
This program simulates a simple vending machine. It displays options for juice or soda and asks the user to enter their choice using numeric input. An if-else statement checks the user’s selection. If the user enters 1, the program dispenses juice. If the user enters 2, it dispenses soda. Any other input is treated as an invalid choice and displays an error message.