Add TemperatureCheck program in Java
This program checks the current temperature in Celsius and provides feedback based on the temperature range.This program asks the user to enter the current temperature in Celsius. It uses a series of if-else statements to categorize the temperature. If the temperature is above 30 degrees, it displays that it is hot. If it is between 20 and 30 degrees, it displays warm. If it is between 10 and 20 degrees, it displays cool. Otherwise, it displays cold.