Create LeapYearCheck.java
This Java program checks whether a given year is a leap year. It takes a year as input from the user and applies conditional statements to test divisibility rules. If the year is divisible by 4 but not by 100, it is a leap year. If divisible by 100, it further checks divisibility by 400 to decide the result.